+++ R/how_many_imputations.R
@@ -1,21 +1,21 @@
 #' How Many Imputations?
 #'
-#' Implements two-stage "how_many_imputations" from von Hippel (2018)
+#' Implements two-stage "how_many_imputations" from von Hippel (2020)
 #'
 #' The old advice of 5-10 imputations is sufficient for a point estimate (e.g.
 #' an estimated coefficient), but not for estimates of standard errors (and
 #' consequently, hypothesis tests or confidence intervals).
 #'
-#' von Hippel (2018) provides a way to calculate the number of imputations
+#' von Hippel (2020) provides a way to calculate the number of imputations
 #' needed to have consistent estimates of the standard error. To do so requires
 #' an estimate of the Fraction of Missing Information (FMI) which can only be
 #' obtained after running some number of imputations. Therefore, von Hippel
-#' (2018) recommends the following procedure:
+#' (2020) recommends the following procedure:
 #'
 #' \enumerate{
 #'   \item Carry out a limited number of imputations to enable estimation of the
-#'   FMI. von Hippel (2018) recommends 20 imputations.
-#'   \item Use this function, \code{howManyImputations}, to calculate how many
+#'   FMI. von Hippel (2020) recommends 20 imputations.
+#'   \item Use this function, \code{how_many_imputations}, to calculate how many
 #'   total imputations you will need.
 #'   \item If the number of total imputations you will need is larger than your
 #'   initial batch of 20, run additional imputations.
@@ -74,4 +74,18 @@ how_many_imputations <- function(model,
   fmiu <- plogis(qlogis(fmi) + z*sqrt(2/model$m))
 
   ceiling(1 + 1/2*(fmiu/cv)^2)
+}
+
+#' Improper function name, use \code{how_many_imputations} instead
+#'
+#' @param model See \code{how_many_imputations}
+#' @param cv See \code{how_many_imputations}
+#' @param alpha See \code{how_many_imputations}
+#' @return See \code{how_many_imputations}
+#' @export
+howManyImputations <- function(model,
+                               cv = .05,
+                               alpha = .05) {
+  warning("Proper function name is how_many_imputations")
+  howManyImputations::how_many_imputations(model, cv, alpha)
 }