+++ NEWS.md
@@ -1,3 +1,11 @@
+# howManyImputations 0.2.2.9001
+
+- Because this package is just a thin wrapper around von Hippel (2020)'s work,
+  added citation to that paper rather than this package.
+- Added function `howManyImputations()` with a warning to direct users to the
+  proper `how_many_imputations()`.
+- Fixes to documentation.
+
 # howManyImputations 0.2.2
 
 Additional metadata tweaks for CRAN.